Como administrador quiero ver la información de los clientes de las compras de mis sesiones
git-svn-id: http://jira.uji.es/repos/uji/PAR/uji-par/trunk@30649 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ uji-par-base/src/test/java/es/uji/apps/par/services/dao/ButacasDAOTest.java
@@ -108,7 +108,7 @@ public class ButacasDAOTest extends BaseDAOTest
         Assert.assertEquals(localizacion.getCodigo(), butacas.get(0).getParLocalizacion().getCodigo());
         Assert.assertEquals(butaca.getFila(), butacas.get(0).getFila());
         Assert.assertEquals(butaca.getNumero(), butacas.get(0).getNumero());
-        Assert.assertEquals(BigDecimal.valueOf(10), butacas.get(0).getPrecio());
+        Assert.assertEquals(10, butacas.get(0).getPrecio().intValue());
     }
     
     @Transactional
